+++ CHANGELOG.md
@@ -22,6 +22,8 @@ None.
 - RNN models deposited on Zenodo and pretrained on ChEMBL31 and Papyrus 05.5 were updated while the RNN model pretrained on ChEMBL27 did not need to.
 - Moved encoding of tokens for SMILES-based models to the parallel preprocessing steps to improve performance
 - All testing code that is not unit tests was moved to `testing`
+- Remove QSAR modelling from DrugEx, now in a seperate repository QSPRpred
+
 
 ## New Features